move user detials functions to kube package
